PageGraph mega commit
It'd be good to do a proper rebase with the full history of the
PageGraph implementation separated into logical commits. Unfortunately,
this is a time consuming process and nobody has that kind of time
available. Also, half the commit history is lost to different rebasing
strategies at the time of writing this message anyways.